Add exception tests
