+++ pypowsybl/utils/dataframes.py
@@ -1,0 +1,101 @@
+#
+# Copyright (c) 2022, RTE (http://www.rte-france.com)
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+"""
+Provides utility methods for dataframes handling:
+ - mapping with keyword arguments
+ - creation of C API dataframes
+ - ...
+"""
+from typing import List
+from pandas import DataFrame, Index, MultiIndex
+from pypowsybl._pypowsybl import SeriesMetadata, create_dataframe
+import numpy as np
+
+
+def _to_array(value):
+    """
+    Converts a scalar or array to an array
+    """
+    as_array = np.array(value, ndmin=1, copy=False)
+    if as_array.ndim != 1:
+        raise ValueError('Network elements update: expecting only scalar or 1 dimension array '
+                         'as keyword argument, got {} dimensions'.format(as_array.ndim))
+    return as_array
+
+
+def _adapt_kwargs(metadata: List[SeriesMetadata], **kwargs) -> DataFrame:
+    """
+    Converts named arguments to a dataframe.
+    """
+    index_columns = [col.name for col in metadata if col.is_index]
+
+    columns = {}
+    expected_size = None
+    for key, value in kwargs.items():
+        col = _to_array(value)
+        size = col.shape[0]
+        if expected_size is None:
+            expected_size = size
+        elif size != expected_size:
+            raise ValueError('Network elements update: all arguments must have the same size, '
+                             'got size {} for series {}, expected {}'.format(size, key, expected_size))
+        columns[key] = col
+
+    index = None
+    if len(index_columns) == 1:
+        index_name = index_columns[0]
+        if not index_name in columns:
+            raise ValueError('No data provided for index: ' + index_name)
+        index = Index(name=index_name, data=columns[index_name])
+    elif len(index_columns) > 1:
+        index = MultiIndex.from_arrays(names=index_columns, arrays=[columns[name] for name in index_columns])
+    data = dict((k, v) for k, v in columns.items() if k not in index_columns)
+    return DataFrame(index=index, data=data)
+
+
+def _adapt_df_or_kwargs(metadata: List[SeriesMetadata], df: DataFrame, **kwargs) -> DataFrame:
+    """
+    Ensures we get a dataframe, either from a ready to use dataframe, or from keyword arguments.
+    """
+    if df is None:
+        return _adapt_kwargs(metadata, **kwargs)
+    elif kwargs:
+        raise RuntimeError('You must provide data in only one form: dataframe or named arguments')
+    return df
+
+
+def _create_c_dataframe(df: DataFrame, series_metadata: List[SeriesMetadata]):
+    """
+    Creates the C representation of a dataframe.
+    """
+    metadata_by_name = {s.name: s for s in series_metadata}
+    is_index = []
+    columns_names = []
+    columns_values = []
+    columns_types = []
+    is_multi_index = len(df.index.names) > 1
+
+    for idx, index_name in enumerate(df.index.names):
+        if index_name is None:
+            index_name = series_metadata[idx].name
+        if is_multi_index:
+            columns_values.append(df.index.get_level_values(index_name))
+        else:
+            columns_values.append(df.index.values)
+        columns_names.append(index_name)
+        columns_types.append(metadata_by_name[index_name].type)
+        is_index.append(True)
+    columns_names.extend(df.columns.values)
+    for series_name in df.columns.values:
+        if not series_name in metadata_by_name:
+            raise ValueError('No column named {}'.format(series_name))
+        series = df[series_name]
+        series_type = metadata_by_name[series_name].type
+        columns_types.append(series_type)
+        columns_values.append(series.values)
+        is_index.append(False)
+    return create_dataframe(columns_values, columns_names, columns_types, is_index)
